news-article-search/requirements.txt
nltk==3.5
oyaml==0.9
docker==4.2.0
newspaper3k==0.2.8
pandas==1.0.3